+++ tests/sim/gridworld/test_observer.py
@@ -32,13 +32,13 @@ def test_single_grid_observer():
     assert observer.supported_agent_type == GridObservingAgent
     assert isinstance(observer, ObserverBaseComponent)
     assert agents['agent0'].observation_space['grid'] == Box(
-        -2, 6, (5, 5), np.int
+        -2, 6, (5, 5), int
     )
     assert agents['agent1'].observation_space['grid'] == Box(
-        -2, 6, (3, 3), np.int
+        -2, 6, (3, 3), int
     )
     assert agents['agent2'].observation_space['grid'] == Box(
-        -2, 6, (9, 9), np.int
+        -2, 6, (9, 9), int
     )
 
     position_state.reset()
@@ -179,13 +179,13 @@ def test_multi_grid_observer():
     assert isinstance(observer, ObserverBaseComponent)
     assert observer.number_of_encodings == 6
     assert agents['agent0'].observation_space['grid'] == Box(
-        -2, 9, (5, 5, 6), np.int
+        -2, 9, (5, 5, 6), int
     )
     assert agents['agent1'].observation_space['grid'] == Box(
-        -2, 9, (3, 3, 6), np.int
+        -2, 9, (3, 3, 6), int
     )
     assert agents['agent2'].observation_space['grid'] == Box(
-        -2, 9, (9, 9, 6), np.int
+        -2, 9, (9, 9, 6), int
     )
     position_state.reset()
